Unit test mock objects moved to __init__.py
